src/requirements.txt
functions-framework==3.*
google-cloud-logging
google-auth
google-api-python-client
vertexai